backend/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pymupdf==1.22.5
pytesseract==0.3.10
